# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stone World API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: origin
          in: query
          schema:
            type: string
            enum: [imported, national, all]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/bulk-delete:
    post:
      operationId: bulkDeleteProducts
      tags: [products]
      summary: Bulk delete products (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteBody"
      responses:
        "200":
          description: Products deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/bulk-toggle:
    post:
      operationId: bulkToggleProductAvailability
      tags: [products]
      summary: Bulk toggle product availability (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkToggleBody"
      responses:
        "200":
          description: Products updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Blog ────────────────────────────────────────────────────────────────────
...
  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries (admin)
      parameters:
        - name: audience
          in: query
          schema:
            type: string
            enum: [homeowner, contractor, architect, developer, all]
        - name: read
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Enquiries list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquiryListResponse"
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit an enquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryBody"
      responses:
        "201":
          description: Enquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /enquiries/{id}/read:
    put:
      operationId: markEnquiryRead
      tags: [enquiries]
      summary: Mark enquiry as read/unread (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadBody"
      responses:
        "200":
          description: Enquiry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
  /enquiries/export:
    get:
      operationId: exportEnquiries
      tags: [enquiries]
      summary: Export enquiries as CSV (admin)
      parameters:
        - name: audience
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # ─── Media ───────────────────────────────────────────────────────────────────
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List all media files (admin)
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaItem"
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file via base64 data URL (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadMediaBody"
      responses:
        "201":
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaItem"
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Settings ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Admin ─────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        priceUnit:
          type: string
          nullable: true
        origin:
          type: string
          enum: [imported, national]
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        deliveryAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - origin
        - available
        - featured
        - deliveryAvailable
        - images
        - tags
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        price:
          type: number
          nullable: true
        priceUnit:
          type: string
          nullable: true
        origin:
          type: string
          enum: [imported, national]
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        deliveryAvailable:
          type: boolean
      required:
        - name
        - origin
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - products
        - total
        - limit
        - offset
    BulkDeleteBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
    BulkToggleBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        available:
          type: boolean
      required:
        - ids
        - available

    # ─── Blog ──────────────────────────────────────────────────────────────────
...
    Enquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
        audience:
          type: string
          enum: [homeowner, contractor, architect, developer]
        projectType:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        message:
          type: string
        productInterest:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - audience
        - message
        - isRead
        - createdAt
    CreateEnquiryBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
        audience:
          type: string
          enum: [homeowner, contractor, architect, developer]
        projectType:
          type: string
          nullable: true
        budget:
          type: string
          nullable: true
        message:
          type: string
        productInterest:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required:
        - name
        - phone
        - audience
        - message
    MarkReadBody:
      type: object
      properties:
        isRead:
          type: boolean
      required:
        - isRead
    EnquiryListResponse:
      type: object
      properties:
        enquiries:
          type: array
          items:
            $ref: "#/components/schemas/Enquiry"
        total:
          type: integer
        unread:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - enquiries
        - total
        - unread
        - limit
        - offset

    # ─── Media ─────────────────────────────────────────────────────────────────
    UploadMediaBody:
      type: object
      properties:
        filename:
          type: string
        dataUrl:
          type: string
        mimeType:
          type: string
      required:
        - filename
        - dataUrl
        - mimeType
    MediaItem:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        url:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - filename
        - url
        - mimeType
        - size
        - createdAt

    # ─── Settings ──────────────────────────────────────────────────────────────
...